extract fargate into a standalone file
